requirements.txt
# Core dependencies
python-dotenv>=1.0.0
dspy-ai>=2.5.0

# LLM Tracking & Monitoring
weave>=0.50.0
wandb>=0.16.0

# LLM Provider SDKs
anthropic>=0.25.0
openai>=1.0.0
google-genai>=0.2.0
mistralai>=0.1.0

# Data handling
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0